Added market_hash_name field
Faced a problem where i need to get item names as in market. The 'name' field will return only skin name, without exterior. So i need to do it on my own, and check if there is exterior, because it can be a sticker. And steam have market_hash_name field, so would be nice to have it.